+++ sync-for-payroll/RELEASES.md
@@ -59,3 +59,13 @@ Based on:
 - [typescript v1.3.0] sync-for-payroll
 ### Releases
 - [NPM v1.3.0] https://www.npmjs.com/package/@codat/sync-for-payroll/v/1.3.0 - sync-for-payroll
+
+## 2023-10-30 01:23:30
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payroll.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.4.0] sync-for-payroll
+### Releases
+- [NPM v1.4.0] https://www.npmjs.com/package/@codat/sync-for-payroll/v/1.4.0 - sync-for-payroll